refactor: 의미없는 파일 삭제
